oprava SQLService a rozdelenie Models do samostatnych suborov
