+++ code/Services/TemplateManagerSystem.cs
@@ -14,14 +14,14 @@ namespace code.Services
 
 		public async void AddTemplate(TrainTemplate tmp)
 		{
-			
 			string sql = "INSERT INTO templates (name, destination, max_lenght) VALUES ((@p1),(@p2),(@p3))";
 			
-			List<NpgsqlParameter> parameters = new List<NpgsqlParameter>();
-			parameters.Add(new NpgsqlParameter("p1", tmp.Name));
-			parameters.Add(new NpgsqlParameter("p2", tmp.Destination));
-			parameters.Add(new NpgsqlParameter("p7", tmp.MaxLength));
-
+			List<NpgsqlParameter> parameters = new List<NpgsqlParameter>
+			{
+				new NpgsqlParameter("p1", tmp.Name),
+				new NpgsqlParameter("p2", tmp.Destination),
+				new NpgsqlParameter("p3", tmp.MaxLength)
+			};
 			
 			NpgsqlDataReader reader = await s.sqlCommand(sql,parameters);
 			reader.Close();
@@ -33,8 +33,6 @@ namespace code.Services
 			string sql = "SELECT * from templates";
 
 			List<NpgsqlParameter> parameters = new List<NpgsqlParameter>();
-			parameters.Add(new NpgsqlParameter("p1", id));
-
 
 			NpgsqlDataReader reader = await s.sqlCommand(sql,parameters);
 
@@ -48,7 +46,7 @@ namespace code.Services
 					Destination = (string)reader[2],	
 					MaxLength=(int)reader[3]
 				};
-                templates.add(tmp);
+                templates.Add(tmp);
 			}
 			reader.Close();
 			return templates;
@@ -58,11 +56,13 @@ namespace code.Services
 		{
 			string sql = "UDPATE templates SET name = (@p2), destination = (@p3), max_lenght = (@p4) WHERE id = (@p1)";
 
-			List<NpgsqlParameter> parameters = new List<NpgsqlParameter>();
-			parameters.Add(new NpgsqlParameter("p1", tmp.Id));
-			parameters.Add(new NpgsqlParameter("p2", tmp.Name));
-			parameters.Add(new NpgsqlParameter("p3", tmp.Destination));
-			parameters.Add(new NpgsqlParameter("p4", tmp.MaxLength));
+			List<NpgsqlParameter> parameters = new List<NpgsqlParameter>()
+			{
+				new NpgsqlParameter("p1", tmp.Id),
+				new NpgsqlParameter("p2", tmp.Name),
+				new NpgsqlParameter("p3", tmp.Destination),
+				new NpgsqlParameter("p4", tmp.MaxLength)
+			};
 
 			NpgsqlDataReader reader = await s.sqlCommand(sql,parameters);
 
@@ -73,8 +73,10 @@ namespace code.Services
 		{
 			string sql = "DELETE FROM templates WHERE id = (@p1)";
 
-			List<NpgsqlParameter> parameters = new List<NpgsqlParameter>();
-			parameters.Add(new NpgsqlParameter("p1", tmp.Id));
+			List<NpgsqlParameter> parameters = new List<NpgsqlParameter>()
+			{
+			new NpgsqlParameter("p1", tmp.Id)
+			};
 
 			NpgsqlDataReader reader = await s.sqlCommand(sql,parameters);
